Move views into a new namespace
