+++ src/AudioBand/Views/Winforms/EnhancedProgressBar.cs
@@ -1,12 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace AudioBand
+namespace AudioBand.Views.Winforms
 {
     // So we can draw with our own color
     public class EnhancedProgressBar : ProgressBar
