+++ src/AudioBand/Views/Wpf/ImagePicker.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="AudioBand.Settings.ImagePicker"
+﻿<UserControl x:Class="AudioBand.Views.Wpf.ImagePicker"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
