+++ src/AudioBand/Views/Wpf/ResetButton.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="AudioBand.Settings.ResetButton"
+﻿<UserControl x:Class="AudioBand.Views.Wpf.ResetButton"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
