+++ src/AudioBand/Views/Wpf/ResetButton.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace AudioBand.Views.Wpf
+{
+    /// <summary>
+    /// Interaction logic for ResetButton.xaml
+    /// </summary>
+    public partial class ResetButton : UserControl
+    {
+        public ResetButton()
+        {
+            InitializeComponent();
+        }
+    }
+}
